feat : change the component tree
- use react fragment
